+++ actionmailbox/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Add Amazon SES/SNS ingress.
+
+    *Bob Farrell*
+